FINAL GESTURE/gesture controlled media player/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
PyMongo==4.6.0
PyJWT==2.8.0
Werkzeug==2.3.7